+++ src/player.rs
@@ -6,12 +6,12 @@ use crate::net::MessageEvent;
 
 #[derive(Serialize)]
 pub(crate) struct Player {
-    pub id: u64,
+    pub id: u32,
     #[serde(skip_serializing)]
     pub tx: Sender<MessageEvent>,
 }
 impl Player {
-    pub fn new(id: u64, tx: Sender<MessageEvent>) -> Self {
+    pub fn new(id: u32, tx: Sender<MessageEvent>) -> Self {
         Self {
             id: id,
             tx: tx,
